requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
boto3==1.28.20
botocore==1.31.20
cfgv==3.3.1
click==8.1.6
colorama==0.4.6
distlib==0.3.7
ecdsa==0.18.0
fastapi==0.101.0
filelock==3.12.2
h11==0.14.0
identify==2.5.26
idna==3.4
jmespath==1.0.1
mangum==0.17.0
nodeenv==1.8.0
passlib==1.7.4
platformdirs==3.10.0
pre-commit==3.3.3
pyasn1==0.5.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
virtualenv==20.24.2
